Add files via upload
The "Titanic Dataset Classification" is a compelling and educational project aimed at building a robust machine learning system in Python to predict whether a passenger on the Titanic would have survived the tragic sinking. This GitHub repository contains a comprehensive implementation of the classification model, analyzing various factors like socio-economic status, age, gender, and more to determine the likelihood of survival for each passenger.